requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
autobahn==24.4.2
Automat==22.10.0
billiard==4.2.0
boto3==1.34.122
botocore==1.34.122
Brotli==1.1.0
cachetools==5.4.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.7
daphne==4.1.2
defusedxml==0.8.0rc2
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.1.0
django-cors-headers==4.3.1
django-dirtyfields==1.9.3
django-filter==24.2
django-mailbox==4.10.1
django-model-utils==4.5.1
django-notifications-hq==1.8.3
django-rest-knox==5.0.1
django-rest-swagger==2.2.0
django-storages==1.14.3
django-templated-mail==1.1.1
djangorestfilemanager==2.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-chunked-upload==0.6.0
drf-yasg==1.21.7
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
gunicorn==22.0.0
h11==0.14.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonfield==3.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.0
pillow==10.4.0
pluggy==1.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
service-identity==24.1.0
setuptools==72.1.0
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
swapper==1.4.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uuid==1.30
uvicorn==0.30.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
zope.interface==7.0.1
coverage==7.4.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
wait-for-it==2.2.2

